requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
sqlmodel
alembic
psycopg2-binary
fastapi-mail
